+++ src/test/merchantRepublic.t.sol
@@ -18,13 +18,10 @@ contract MRTest is Gov2Test {
         assertEq(remain2, john.silverBalance());
     }
 
-
-
-
     function testGuildsBlockProposalVote() public {
         initCommoners();
         createProposalTarget();
-        uint voteStartDay = block.timestamp + 2.5 days;
+        uint256 voteStartDay = block.timestamp + 2.5 days;
         hevm.warp(voteStartDay);
         uint48 guildId = 0;
         uint8 support = 0;
@@ -63,7 +60,6 @@ contract MRTest is Gov2Test {
         );
     }
 
-
     function testPassProposalVote() public {
         initCommoners();
         createProposalTarget();
@@ -98,7 +94,8 @@ contract MRTest is Gov2Test {
 
         // Guild Vote
         ursus.guildCastVoteForProposal(support, id, guildId);
-        (uint startTimestamp, uint endTimestamp) = merchantRepublic.getTimes(id);
+        (uint256 startTimestamp, uint256 endTimestamp) = merchantRepublic
+            .getTimes(id);
         emit log_named_uint("Guild Verdict returned: ", block.timestamp);
         emit log_named_uint("Vote startTimestamp: ", startTimestamp);
         emit log_named_uint("Vote endTimestamp: ", endTimestamp);
@@ -116,7 +113,7 @@ contract MRTest is Gov2Test {
             uint256(MerchantRepublic.ProposalState.PendingCommonersVoteStart)
         );
         emit log_named_uint("Vote start: ", block.timestamp);
-        uint voteStartDay = block.timestamp + 2 days + 1;
+        uint256 voteStartDay = block.timestamp + 2 days + 1;
         hevm.warp(voteStartDay);
         emit log_named_uint("Commoners vote: ", block.timestamp);
         // The voting delay has passed and the proposal is ready to be
@@ -132,7 +129,7 @@ contract MRTest is Gov2Test {
         // by 2.5 days, so we arrived at the middle of the first day of voting.
         // Thus, we only need to warp 5.5 days into the future for the vote to end.
         // We warp 6 days into the future for good measure
-        uint voteEndDay = block.timestamp + 7 days;
+        uint256 voteEndDay = block.timestamp + 7 days;
         hevm.warp(voteEndDay);
         emit log_named_uint("Proposal Queued: ", block.timestamp);
         assertEq(
@@ -149,6 +146,7 @@ contract MRTest is Gov2Test {
             uint256(merchantRepublic.state(id))
         );
     }
+
     function testGuildsVoteOnWrongProposal() public {
         initCommoners();
         hevm.warp(block.timestamp + 7 days);
@@ -186,20 +184,17 @@ contract MRTest is Gov2Test {
     }
 
     function testSetMerchantRepublicParameters() public {
-       // from the Gov2Test contract, during the setUp()
-       // function, ursus is the creator of the Merchant
-       // Republic smart contract, and thus the doge.
-       ursus.govSetVotingDelay(10 days);
-       ursus.govSetVotingPeriod(20 days);
-       ursus.govSetProposalThreshold(1009999e19);
-       ursus.govSetPendingDoge(address(agnello));
-       agnello.govAcceptDoge();
-       assertEq(10 days, merchantRepublic.votingDelay());
-       assertEq(20 days, merchantRepublic.votingPeriod());
-       assertEq(1009999e19, merchantRepublic.proposalThreshold());
-       assertEq(address(agnello), merchantRepublic.doge());
+        // from the Gov2Test contract, during the setUp()
+        // function, ursus is the creator of the Merchant
+        // Republic smart contract, and thus the doge.
+        ursus.govSetVotingDelay(10 days);
+        ursus.govSetVotingPeriod(20 days);
+        ursus.govSetProposalThreshold(1009999e19);
+        ursus.govSetPendingDoge(address(agnello));
+        agnello.govAcceptDoge();
+        assertEq(10 days, merchantRepublic.votingDelay());
+        assertEq(20 days, merchantRepublic.votingPeriod());
+        assertEq(1009999e19, merchantRepublic.proposalThreshold());
+        assertEq(address(agnello), merchantRepublic.doge());
     }
-
-
-
 }
